refactor(api): move auth-status / status / overview / costs / latency / logs to server/api/dashboard/
Commit 2 of the dashboard fold-in. Moves six read-only endpoints out
of dashboard/api/main.py into three new files under
server/api/dashboard/:

- auth_status.py: GET /api/auth-status
- status.py:      GET /api/status, GET /api/overview
- costs.py:       GET /api/costs, GET /api/latency, GET /api/logs

Each handler swaps the module-global ._gateway lookup for
Depends(get_gateway) from server/api/_deps.py, matching how every
/v1/* router resolves the gateway today. Behavior is preserved
verbatim: query schemas, return shapes, the _LOCAL_PROVIDER_NAMES
frozenset, the pricing-source subtree, the storage-None fallback,
all carry over. The new /api/costs lives alongside the existing
server/api/costs.py (which serves /v1/costs with a different query
shape).

tests/server/test_dashboard_api.py now points its ASGITransport at
the daemon's build_app(gateway) instead of dashboard.api.main.app.
Until commit 7, both halves remain reachable through the daemon:
moved endpoints serve via the new dashboard_router; unmoved endpoints
serve via ApplicationBuilder._mount_dashboard's runtime route graft.

ruff check src + uv run mypy + uv run pytest -q all green: 1666 passed,
4 skipped, same total as before the refactor.
