+++ src/dashboard/api/main.py
@@ -15,7 +15,7 @@ from fastapi.staticfiles import StaticFiles
 from pydantic import ValidationError
 from sqlalchemy import text
 
-from voicegateway.core.auth import load_api_keys, resolve_cors_origins
+from voicegateway.core.auth import resolve_cors_origins
 from voicegateway.repository import (
     dead_air_repository as dead_air,
 )
@@ -118,174 +118,6 @@ def configure(gateway: Gateway) -> None:
         allow_headers=["*"],
     )
     _cors_configured = True
-
-
-@app.get("/api/auth-status")
-async def get_auth_status() -> dict:
-    """Report whether the gateway requires a token for mutating endpoints.
-
-    Used by the frontend to decide whether to show the login gate. The
-    response never exposes key material.
-    """
-    gw = _get_gateway()
-    keys = load_api_keys(gw.config.auth)
-    return {"auth_required": bool(keys)}
-
-
-@app.get("/api/status")
-async def get_status() -> dict:
-    """Get status of all configured providers and models.
-
-    Mirrors the pricing-freshness subtree from /v1/status (Q7) so
-    the dashboard StalenessBanner can render without hitting a
-    second origin.
-    """
-    gw = _get_gateway()
-    config = gw.config
-
-    providers = {}
-    for name, cfg in config.providers.items():
-        is_local = name in _LOCAL_PROVIDER_NAMES
-        providers[name] = {
-            "configured": bool(cfg.get("api_key")) or is_local,
-            "type": "local" if is_local else "cloud",
-        }
-
-    models = {}
-    for modality, modality_models in config.models.items():
-        if isinstance(modality_models, dict):
-            for model_id, model_cfg in modality_models.items():
-                if isinstance(model_cfg, dict):
-                    models[model_id] = {
-                        "modality": modality,
-                        "provider": model_cfg.get("provider", ""),
-                    }
-
-    from voicegateway.inference.pricing import llm as _llm_pricing
-    from voicegateway.inference.pricing import stt as _stt_pricing
-    from voicegateway.inference.pricing import tts as _tts_pricing
-
-    pricing = {
-        "llm": {"source": _llm_pricing.PRICING_SOURCE},
-        "stt": {
-            "source": _stt_pricing.PRICING_SOURCE,
-            "oldest_entry_date": _stt_pricing._oldest_pricing_date().isoformat(),
-        },
-        "tts": {
-            "source": _tts_pricing.PRICING_SOURCE,
-            "oldest_entry_date": _tts_pricing._oldest_pricing_date().isoformat(),
-        },
-    }
-
-    return {
-        "providers": providers,
-        "models": models,
-        "fallbacks": config.fallbacks,
-        "pricing": pricing,
-    }
-
-
-@app.get("/api/costs")
-async def get_costs(
-    period: str = Query("today", enum=["today", "week", "month", "all"]),
-    project: str | None = Query(None),
-    tenant: str | None = Query(None),
-) -> dict:
-    """Get cost summary for a period, optionally filtered by project and tenant.
-
-    Each ``by_model`` entry carries a ``pricing_source`` string (e.g.
-    ``"genai-prices@0.0.57"`` or ``"voicegateway-catalog@2026-05-04"``)
-    so the dashboard can render the cost-staleness banner without a
-    second round-trip. ``tenant`` accepts a tenant id; pass the empty
-    string to scope to the unattributed bucket (REQ-VG-TENANT-002).
-    """
-    gw = _get_gateway()
-    if gw.storage is None:
-        return {
-            "period": period,
-            "project": project,
-            "total": 0.0,
-            "by_provider": {},
-            "by_model": {},
-            "by_project": {},
-        }
-    summary = await gw.storage.get_cost_summary(
-        period, project=project, include_pricing_source=True, tenant=tenant
-    )
-    if project is None:
-        summary["by_project"] = await gw.storage.get_cost_by_project(
-            period, tenant=tenant
-        )
-    else:
-        summary["by_project"] = {}
-    return summary
-
-
-@app.get("/api/latency")
-async def get_latency(
-    period: str = Query("today", enum=["today", "week"]),
-    project: str | None = Query(None),
-    tenant: str | None = Query(None),
-) -> dict:
-    """Get latency statistics, optionally filtered by project and tenant."""
-    gw = _get_gateway()
-    if gw.storage is None:
-        return {}
-    pcts = gw.config.latency.get("percentiles") or [50.0, 95.0, 99.0]
-    return await gw.storage.get_latency_stats(
-        period, project=project, percentiles=pcts, tenant=tenant
-    )
-
-
-@app.get("/api/logs")
-async def get_logs(
-    limit: int = Query(100, ge=1, le=1000),
-    modality: str | None = Query(None, enum=["stt", "llm", "tts"]),
-    project: str | None = Query(None),
-    tenant: str | None = Query(None),
-) -> list[dict]:
-    """Get recent request logs, optionally filtered by modality, project, and tenant."""
-    gw = _get_gateway()
-    if gw.storage is None:
-        return []
-    return await gw.storage.get_recent_requests(
-        limit=limit, modality=modality, project=project, tenant=tenant
-    )
-
-
-@app.get("/api/overview")
-async def get_overview(
-    project: str | None = Query(None),
-) -> dict:
-    """Get dashboard overview stats, optionally filtered by project."""
-    gw = _get_gateway()
-    config = gw.config
-
-    model_count = 0
-    for modality_models in config.models.values():
-        if isinstance(modality_models, dict):
-            model_count += len(modality_models)
-
-    if gw.storage is None:
-        return {
-            "total_requests": 0,
-            "total_cost": 0.0,
-            "active_models": model_count,
-            "providers_configured": len(config.providers),
-        }
-
-    cost_summary = await gw.storage.get_cost_summary("today", project=project)
-    total_all = await gw.storage.get_cost_summary("all", project=project)
-
-    return {
-        "total_requests": sum(
-            d["requests"] for d in total_all.get("by_provider", {}).values()
-        ),
-        "total_cost_today": cost_summary.get("total", 0.0),
-        "total_cost_all": total_all.get("total", 0.0),
-        "active_models": model_count,
-        "providers_configured": len(config.providers),
-    }
 
 
 @app.get("/api/projects")
